requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.4.2
motor==3.3.1
python-jose==3.3.0 
pydantic[email]==2.4.2